requirements.txt
oauthlib==0.7.2
python-twitter==2.2
requests==2.6.0
requests-oauthlib==0.4.2
